+++ IBSWeb/Areas/User/Controllers/CustomerController.cs
@@ -189,6 +189,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> GetCustomersList([FromForm] DataTablesParameters parameters, CancellationToken cancellationToken)
         {
             try
@@ -267,6 +268,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost, ActionName("Activate")]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> ActivatePost(int? id, CancellationToken cancellationToken)
         {
             if (id == null || id == 0)
@@ -340,6 +342,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost, ActionName("Deactivate")]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeactivatePost(int? id, CancellationToken cancellationToken)
         {
             if (id == null || id == 0)
@@ -518,6 +521,7 @@ namespace IBSWeb.Areas.User.Controllers
         #region -- export xlsx record --
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Export(string selectedRecord)
         {
             if (string.IsNullOrEmpty(selectedRecord))