+++ IBSWeb/Areas/User/Controllers/PaymentTermsController.cs
@@ -44,6 +44,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> GetPaymentTerms([FromForm] DataTablesParameters parameters, CancellationToken cancellationToken)
         {
             try
@@ -210,6 +211,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Delete(string code, CancellationToken cancellationToken)
         {
             if (string.IsNullOrEmpty(code))