+++ IBSWeb/Areas/User/Controllers/PortController.cs
@@ -35,6 +35,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Port model, CancellationToken cancellationToken = default)
         {
             var result = await portService.CreateAsync(model, userManager.GetUserName(User)!, cancellationToken);
@@ -61,6 +62,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Port model, CancellationToken cancellationToken)
         {
             var result = await portService.UpdateAsync(model, userManager.GetUserName(User)!, cancellationToken);