+++ IBSWeb/Areas/User/Controllers/ServiceRequestController.cs
@@ -434,6 +434,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> GetDispatchTicketLists([FromForm] DataTablesParameters parameters, CancellationToken cancellationToken)
         {
             var currentUser = await userManager.GetUserAsync(User);