+++ IBSWeb/Areas/User/Controllers/VesselScheduleController.cs
@@ -66,6 +66,7 @@ namespace IBSWeb.Areas.User.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public async Task<IActionResult> GetScheduleList([FromForm] DataTablesParameters parameters, DateTime? dateFilter, CancellationToken ct)
         {
             var from = (dateFilter ?? DateTimeHelper.GetCurrentPhilippineTime()).Date;